requirements.txt
streamlit
pandas
openpyxl
boto3
python-dotenv
pydantic